backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=2.0
django-cors-headers>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
celery>=5.4
redis>=5.0
